+++ src/Pim/Bundle/EnrichBundle/MassEditAction/Operation/Classify.php
@@ -2,6 +2,7 @@
 
 namespace Pim\Bundle\EnrichBundle\MassEditAction\Operation;
 
+use Akeneo\Component\Persistence\BulkSaverInterface;
 use Pim\Bundle\CatalogBundle\Manager\CategoryManager;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 
@@ -14,9 +15,7 @@ use Pim\Bundle\CatalogBundle\Model\ProductInterface;
  */
 class Classify extends ProductMassEditOperation
 {
-    /**
-     * @var CategoryManager $categoryManager
-     */
+    /** @var CategoryManager $categoryManager */
     protected $categoryManager;
 
     /**
@@ -30,10 +29,12 @@ class Classify extends ProductMassEditOperation
     protected $categories;
 
     /**
-     * @param CategoryManager $categoryManager
+     * @param CategoryManager    $categoryManager
+     * @param BulkSaverInterface $productSaver
      */
-    public function __construct(CategoryManager $categoryManager)
+    public function __construct(CategoryManager $categoryManager, BulkSaverInterface $productSaver)
     {
+        parent::__construct($productSaver);
         $this->categoryManager = $categoryManager;
         $this->trees           = $categoryManager->getEntityRepository()->findBy(array('parent' => null));
         $this->categories      = array();